feat: use cdn images for map pin calls
